Factor out NormalPicker
